refactor: extract sanitizeCommandLineArgs into shared security-utils module
- Create scripts/security-utils.ts with sanitizeCommandLineArgs function
- Update mcp-launcher.ts to import from shared module instead of duplicating
- Update mcp-launcher.test.ts to test actual implementation via import
- Remove function duplication that risked test/implementation divergence
- Add logPrefix parameter to shared function for consistent logging
- Ensure tests validate real code behavior instead of duplicated logic

This ensures tests validate the actual implementation and prevents
maintenance issues from keeping duplicate code in sync.
