+++ scripts/mcp-launcher.ts
@@ -17,6 +17,7 @@ import { constants } from 'node:fs'
 import { access, readdir, watch } from 'node:fs/promises'
 import path from 'node:path'
 import { fileURLToPath } from 'node:url'
+import { sanitizeCommandLineArgs } from './security-utils.ts'
 
 // --- Constants ---
 const colors = {
@@ -49,62 +50,7 @@ const sessionId = Math.random().toString(36).substring(2, 8)
 process.env.PM_SESSION_ID = sessionId
 
 // Forward all command line arguments to the MCP server with validation
-const forwardedArgs = sanitizeCommandLineArgs(process.argv.slice(2))
-
-/**
- * Sanitize command line arguments to prevent shell injection attacks
- * Only allows safe arguments that don't contain shell metacharacters
- */
-function sanitizeCommandLineArgs(args: string[]): string[] {
-  const safeArgs: string[] = []
-
-  // Limit total number of arguments to prevent resource exhaustion
-  if (args.length > 50) {
-    console.warn(
-      `${HOT_RELOAD_PREFIX} Too many arguments provided (${args.length}), limiting to first 50`
-    )
-    args = args.slice(0, 50)
-  }
-
-  for (const arg of args) {
-    // Check for potentially dangerous characters that could be used for shell injection
-    if (typeof arg !== 'string') {
-      console.warn(`${HOT_RELOAD_PREFIX} Skipping non-string argument:`, arg)
-      continue
-    }
-
-    // Reject empty arguments
-    if (arg.length === 0) {
-      continue
-    }
-
-    // Allow only safe characters: alphanumeric, hyphens, underscores, dots, forward slashes, colons, equals
-    // This covers most legitimate MCP server arguments while blocking shell metacharacters
-    // Specifically blocks: $, `, ;, |, &, >, <, (, ), [, ], {, }, *, ?, \, ", ', space, tab
-    if (!/^[a-zA-Z0-9\-_./:=]+$/.test(arg)) {
-      console.warn(`${HOT_RELOAD_PREFIX} Skipping potentially unsafe argument: ${arg}`)
-      continue
-    }
-
-    // Limit argument length to prevent buffer overflow attempts
-    if (arg.length > 500) {
-      console.warn(`${HOT_RELOAD_PREFIX} Skipping oversized argument (${arg.length} chars)`)
-      continue
-    }
-
-    // Additional validation: reject arguments that look like command injection attempts
-    const suspiciousPatterns = [/&&/, /\|\|/, /;/, /\$\(/, /`/, /\$\{/, />/, /</, /\|/]
-
-    if (suspiciousPatterns.some(pattern => pattern.test(arg))) {
-      console.warn(`${HOT_RELOAD_PREFIX} Skipping argument with suspicious pattern: ${arg}`)
-      continue
-    }
-
-    safeArgs.push(arg)
-  }
-
-  return safeArgs
-}
+const forwardedArgs = sanitizeCommandLineArgs(process.argv.slice(2), HOT_RELOAD_PREFIX)
 
 // --- Main Logic ---
 runWithHotReload().catch(error => {
